refactor: 유틸 함수 분리
