+++ src/app/(route)/main/_components/rightNews/MainRightBar.tsx
@@ -45,8 +45,6 @@ const MainRightBar = ({ isDesktop }: MainRightBarProps) => {
     refetch: refetchNewsData,
   } = useGetMainRightBarNewsData({ page: currentPage, size });
 
-  console.log("filteredNewsData: ", filteredNewsData);
-
   // 탭 유틸 함수
   const tap = getRightBarTabs(setButtonActive, buttonActive);
